More endgame knowledge

+++ src/backend/Position.hpp
@@ -233,7 +233,7 @@ public:
 
     // convert position to neural network input features (active indices list)
     // returns number of active features
-    uint32_t ToFeaturesVector(uint32_t* outFeatures) const;
+    uint32_t ToFeaturesVector(uint16_t* outFeatures) const;
 
     void GeneratePawnMoveList(MoveList& outMoveList, uint32_t flags = 0) const;
     void GenerateKnightMoveList(MoveList& outMoveList, uint32_t flags = 0) const;
